Create formatter utility and move formatting methods to it.
