+++ src/htdocs/js/geoserve/NearbyCitiesView.js
@@ -1,6 +1,8 @@
 'use strict';
 
-var View = require('mvc/View'),
+var Format = require('geoserve/Formatter'),
+
+    View = require('mvc/View'),
 
     Util = require('util/Util');
 
@@ -49,16 +51,6 @@ var NearbyCitiesView = function (params) {
     _this.render();
   };
 
-  /**
-   * Convert azimuth in degree's into compass points.
-   */
-  _this._compasswinds = function(azimuth) {
-    var fullwind = 22.5,
-        directions = ['N', 'NNE', 'NE', 'ENE', 'E', 'ESE', 'SE', 'SSE',
-            'S', 'SSW', 'SW', 'WSW', 'W', 'WNW', 'NW', 'NNW', 'N'];
-
-    return directions[(Math.round((azimuth%360)/fullwind))];
-  };
 
   /**
    * Destroy all the things.
@@ -70,22 +62,6 @@ var NearbyCitiesView = function (params) {
     _initialize = null;
     _this = null;
   }, _this.destroy);
-
-  /**
-   * Convert Kilometers to miles
-   */
-  _this._kmToMi = function (km) {
-    return (km * 0.621371);
-  };
-
-  /**
-   * Put commas into a number for display.
-   */
-  _this._numberWithCommas = function (x) {
-    var parts = x.toString().split('.');
-    parts[0] = parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-    return parts.join('.');
-  };
 
   /**
    * Update map to display current location.
@@ -111,10 +87,11 @@ var NearbyCitiesView = function (params) {
             '<span class="city-distance">' +
               properties.distance +
               'km (' +
-              Math.round(this._kmToMi(properties.distance)) + 'mi) ' +
+                Math.round(Format.kilometersmToMiles(properties.distance)) +
+              'mi) ' +
             '</span>' +
             '<span class="direction">' +
-              _this._compasswinds(properties.azimuth) + ' of ' +
+              Format.compassWinds(properties.azimuth) + ' of ' +
             '</span>' +
             '<span class="city-name">' +
               properties.name +
@@ -126,7 +103,8 @@ var NearbyCitiesView = function (params) {
               ', ' + properties.country_name +
             '</span>' +
             '<span class="population">' +
-              ' population ' +_this._numberWithCommas(properties.population) +
+              ' population ' +
+              Format.numberWithCommas(properties.population) +
             '</span>' +
           '</li>'
         );
